UTs on event handlers
git-svn-id: https://svn.code.sf.net/p/quex/code/trunk@4957 db9deaaf-66c7-40b0-811a-ad7aae109365

+++ quex/output/analyzer/modes.py
@@ -111,9 +111,9 @@ $on_buffer_overflow(void*  me /* 'aux' -> 'self' via 'me' */,
 {
     const QUEX_TYPE_LEXATOM* LexemeBegin = self.buffer._lexeme_start_p;
     const QUEX_TYPE_LEXATOM* LexemeEnd   = self.buffer._read_p;
-    const size_t             BufferSize  =   &self.buffer._memory._back[-1] 
-                                           - &self.buffer._memory._front[1] 
-                                           - (ptrdiff_t)(QUEX_SETTING_BUFFER_MIN_FALLBACK_N);
+    const size_t             BufferSize  = (size_t)(  &self.buffer._memory._back[-1] 
+                                                    - &self.buffer._memory._front[1] 
+                                                    - (ptrdiff_t)(QUEX_SETTING_BUFFER_MIN_FALLBACK_N));
     (void)me; (void)LexemeBegin; (void)LexemeEnd; (void)BufferSize;
 $$ON_BUFFER_OVERFLOW$$
 }
